longdouble_to_double

+++ point.h
@@ -1,7 +1,7 @@
 struct Point{
-    long double x,y;
+    double x,y;
     int color;
-    Point(long double _x=0,long double _y=0,int _color=0){
+    Point(double _x=0,double _y=0,int _color=0){
         x=_x;
         y=_y;
         color=_color;
